LayoutTests/imported/w3c/web-platform-tests/web-animations/timing-model/timelines/update-and-send-events-expected.txt


PASS Fires cancel event before requestAnimationFrame 
PASS Fires finish event before requestAnimationFrame 
FAIL Sorts finish events by composite order assert_array_equals: finish events for various animation type should be sorted by composite order property 0, expected "CSSTransition:finish" but got "ScriptAnimation:finish"
FAIL Sorts cancel events by composite order assert_array_equals: cancel events should be sorted by composite order property 0, expected "CSSTransition:cancel" but got "ScriptAnimation:cancel"
PASS Queues a cancel event in transitionstart event callback 
PASS Sorts events for the same transition 
PASS Playback events with the same timeline retain the order in which they arequeued 
FAIL All timelines are updated before running microtasks promise_test: Unhandled rejection with value: object "AbortError: The operation was aborted."
